+++ crates/repos/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "repos"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.86"
+serde = { version = "1.0.208", features = ["derive"] }
+serde_json = "1.0.125"
+sqlx = { version = "0.8.1", features = [
+    "postgres",
+    "runtime-tokio-native-tls",
+    "chrono",
+    "uuid",
+] }
+tokio = { version = "1.39.3", features = [
+    "macros",
+    "process",
+    "rt-multi-thread",
+] }
+uuid = { version = "1.10.0", features = ["v4", "serde"] }
+chrono = "0.4.38"
+async-trait = "0.1.83"
