+++ crates/repos/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod entity;
+pub mod postgres;
+pub mod repo;
