+++ crates/worker/src/infra/mod.rs
@@ -1,7 +1,5 @@
-pub mod episode_repo;
 pub mod ffmpeg;
 pub mod r2_storage;
-pub mod script_repo;
 pub mod task_repo;
 pub mod workdir;
